fmt: store 32-bit version info in S3M and IT files
We don't have to edit the IT code at all, since it already saved
the reserved field as a 32-bit integer.

S3M splits the number between the high 2 and low 2 bytes of the
reserved field. The high 2 come after the edit timer, and the low 2 come
after the edit timer. This is so we can retain compatibility
with S3M files made with git versions of schism, as well as
keeping players and other trackers that know about the edit
timer from having to account for the value moving.

This results in the version info overflow being moved from 5 April 2189
to 19 November 11761230, both of which are absolutely ridiculous.

+++ include/version.h
@@ -29,7 +29,7 @@
 extern const char *ver_short_copyright;
 extern const char *ver_short_based_on;
 extern uint16_t ver_cwtv; /* lower 12 bits of the IT/S3M cwtv field */
-extern uint16_t ver_reserved; /* full version number in case 12 bits are not enough */
+extern uint32_t ver_reserved; /* full version number in case 12 bits are not enough */
 
 SCHISM_PURE extern const char *schism_banner(int classic);
